+++ www/src/pages/plugins.js
@@ -18,7 +18,7 @@ class Plugins extends Component {
               display: `flex`,
               minHeight: `calc(100vh - (${presets.headerHeight} + ${
                 presets.bannerHeight
-              } - 1px))`,
+              }))`,
             }}
           >
             <div
@@ -60,7 +60,7 @@ class Plugins extends Component {
                 }}
               >
                 Please use the search bar to find plugins that will make your
-                blazing fast site even more awesome. If you'd like to create
+                blazing fast site even more awesome. If you{`'`}d like to create
                 your own plugin, see the
                 {` `}
                 <Link to="/docs/plugin-authoring/">Plugin Authoring</Link> page