+++ packages/gatsby-source-wordpress/CHANGELOG.md
@@ -3,6 +3,48 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="3.0.0-rc.8"></a>
+
+# [3.0.0-rc.8](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-wordpress/compare/gatsby-source-wordpress@3.0.0-rc.7...gatsby-source-wordpress@3.0.0-rc.8) (2018-09-13)
+
+**Note:** Version bump only for package gatsby-source-wordpress
+
+<a name="3.0.0-rc.7"></a>
+
+# [3.0.0-rc.7](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-wordpress/compare/gatsby-source-wordpress@3.0.0-rc.6...gatsby-source-wordpress@3.0.0-rc.7) (2018-09-12)
+
+**Note:** Version bump only for package gatsby-source-wordpress
+
+<a name="3.0.0-rc.6"></a>
+
+# [3.0.0-rc.6](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-wordpress/compare/gatsby-source-wordpress@3.0.0-rc.5...gatsby-source-wordpress@3.0.0-rc.6) (2018-09-11)
+
+**Note:** Version bump only for package gatsby-source-wordpress
+
+<a name="3.0.0-rc.5"></a>
+
+# [3.0.0-rc.5](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-wordpress/compare/gatsby-source-wordpress@3.0.0-rc.4...gatsby-source-wordpress@3.0.0-rc.5) (2018-09-11)
+
+**Note:** Version bump only for package gatsby-source-wordpress
+
+<a name="3.0.0-rc.4"></a>
+
+# [3.0.0-rc.4](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-wordpress/compare/gatsby-source-wordpress@3.0.0-rc.3...gatsby-source-wordpress@3.0.0-rc.4) (2018-09-11)
+
+**Note:** Version bump only for package gatsby-source-wordpress
+
+<a name="3.0.0-rc.3"></a>
+
+# [3.0.0-rc.3](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-wordpress/compare/gatsby-source-wordpress@3.0.0-rc.2...gatsby-source-wordpress@3.0.0-rc.3) (2018-09-11)
+
+**Note:** Version bump only for package gatsby-source-wordpress
+
+<a name="3.0.0-rc.2"></a>
+
+# [3.0.0-rc.2](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-wordpress/compare/gatsby-source-wordpress@3.0.0-rc.1...gatsby-source-wordpress@3.0.0-rc.2) (2018-09-07)
+
+**Note:** Version bump only for package gatsby-source-wordpress
+
 <a name="3.0.0-rc.1"></a>
 
 # [3.0.0-rc.1](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-wordpress/compare/gatsby-source-wordpress@3.0.0-rc.0...gatsby-source-wordpress@3.0.0-rc.1) (2018-08-29)